# src/main/resources/openapi.yaml
openapi: "3.0.3"
info:
  title: "Car Logbook Service REST API"
  version: "@project.version@"
security:
  - ApiKeyAuth: [ ]
paths:
  /api/v1/vehicle:
    get:
      summary: Protected vehicle endpoint
      tags:
        - vehicle
      operationId: getVehicles
      security:
        - ApiKeyAuth: [ ]
        - UserServiceToken: [ ]
      responses:
        '200':
          description: Successfully received logbook entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vehicle'
        '400':
          description: Bad Request (Invalid Query Parameters)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server Error (Failed to fetch logs)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a vehicle
      tags:
        - vehicle
      operationId: createVehicle
      security:
        - ApiKeyAuth: [ ]
        - UserServiceToken: [ ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vehicle'
          "application/json; utf-8":
            schema:
              $ref: '#/components/schemas/Vehicle'
          "application/json; charset=utf-8":
            schema:
              $ref: '#/components/schemas/Vehicle'
      responses:
        '200':
          description: Successfully created vehicle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          description: Bad Request (Invalid Query Parameters)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server Error (Failed to fetch logs)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/vehicle/{vehicleId}/logbook:
    get:
      summary: Protected logbook endpoint
      tags:
        - logbook
      operationId: getEntries
      security:
        - ApiKeyAuth: [ ]
        - UserServiceToken: [ ]
      parameters:
        - in: path
          name: vehicleId
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '200':
          description: Successfully received logbook entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LogbookEntry'
        '400':
          description: Bad Request (Invalid Query Parameters)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server Error (Failed to fetch logs)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create Logbook Entry for Vehicle
      tags:
        - logbook
      operationId: createEntry
      security:
        - ApiKeyAuth: [ ]
        - UserServiceToken: [ ]
      parameters:
        - in: path
          name: vehicleId
          schema:
            type: string
            format: uuid
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogbookEntry'
          "application/json; utf-8":
            schema:
              $ref: '#/components/schemas/LogbookEntry'
          "application/json; charset=utf-8":
            schema:
              $ref: '#/components/schemas/LogbookEntry'
      responses:
        '200':
          description: Successfully received logbook entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogbookEntry'
        '400':
          description: Bad Request (Invalid Query Parameters)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server Error (Failed to fetch logs)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/ping:
    get:
      summary: Protected ping endpoint
      tags:
        - ping
      operationId: ping
      security:
        - ApiKeyAuth: [ ]
        - UserServiceToken: [ ]
      responses:
        '200':
          description: Successfully pinged endpoint
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PingResponse'
        '400':
          description: Bad Request (Invalid Query Parameters)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server Error (Failed to fetch logs)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    Vehicle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        make:
          type: string
        model:
          type: string
        licensePlate:
          type: string
        yearOfConstruction:
          type: string
          format: date
        power:
          type: object
          properties:
            hp:
              type: integer
            kw:
              type: integer
        gasoline:
          type: string
          enum:
            - GASOLINE
            - DIESEL
            - HYBRID
            - ELECTRIC
        photo:
          type: string
        primary:
          type: boolean
    LogbookEntry:
      type: object
      properties:
        vehicle:
          type: string
          format: uuid
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        kmStart:
          type: integer
        kmEnd:
          type: integer
        trip:
          type: number
          format: double
        route:
          type: string
        comment:
          type: string
    PingResponse:
      type: object
      properties:
        text:
          type: string
        username:
          type: string
        uid:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        errorCode:
          type: string
          description: the error code
        message:
          type: string
          description: the errors message
    Id:
      description: 'ID'
      type: string
      format: uuid
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
    UserServiceToken:
      description: 'Must contain an encrypted User Service Token.'
      type: http
      scheme: bearer